Fix Windows build: silence size_t->unsigned long narrowing warning (C4267)
send_jpeg_chunked takes jpeg_size as unsigned long (32-bit on Windows), but
jpeg_buf.size() is size_t (64-bit on x64), and MSVC's warning-as-error
build flags this narrowing conversion. Frame sizes never approach 4GB, so
an explicit cast is safe.

+++ src/udp_stream_filter.cpp
@@ -309,7 +309,7 @@ static void encode_thread_func(udp_stream_filter *f)
 		std::vector<int> encode_params = {cv::IMWRITE_JPEG_QUALITY, f->jpeg_quality};
 
 		if (cv::imencode(".jpg", frame, jpeg_buf, encode_params) && !jpeg_buf.empty()) {
-			send_jpeg_chunked(f, jpeg_buf.data(), jpeg_buf.size(), frame_id);
+			send_jpeg_chunked(f, jpeg_buf.data(), (unsigned long)jpeg_buf.size(), frame_id);
 			f->frames_sent++;
 		} else {
 			blog(LOG_WARNING, "[xudp] JPEG compression failed");
